Remove "vpc" parameter
The default value for this parameter is true anyway, so removing it
does not change anything.  The parameter changed its name to "domain"
in version 5 of the AWS Terraform provider, although "vpc" was still
accepted as an alias for backward compatibility.  In version 6 of the
AWS Terraform provider "vpc" is no longer a valid parameter.